made some separation of concerns in files
