+++ canvas-movement-practice/server/src/index.ts
@@ -1,11 +1,12 @@
-import { ClientToServerEvents, ServerToClientEvents, InterServerEvents, SocketData, PlayerStats, BuffType, BuffKey, MapType, keyControlsType, BuffList, InnerBuffProps, PlayerType, innerPropTypes, BuffProps } from "./types.ts";
 import path from "path";
 import { createServer } from "http";
 import express, { Express, Request, Response } from "express";
 import cors from "cors";
 import { Server } from "socket.io";
 
-import { isAttackColiding, isSquareColiding, parseCsvMap, isColidingWithEnvironment, generateRespawnCoords, merge } from "./utility.ts";
+import { ClientToServerEvents, ServerToClientEvents, InterServerEvents, SocketData, PlayerStats, BuffKey, keyControlsType, BuffList, PlayerType, BulletType, BuffObjectType } from "./types.ts";
+import { MAP, OBSTACLES, TICK_RATE, TILE_SIZE, BUFFS } from "./constants.ts";
+import { isAttackColiding, isSquareColiding, isColidingWithEnvironment, generateRespawnCoords, spawnRandomBuff } from "./utility.ts";
 
 const app: Express = express();
 const httpServer = createServer(app);
@@ -27,53 +28,7 @@ app.get("/", (req: Request, res: Response) => {
     res.send("Home to TS converted server");
 });
 
-const MAP: MapType = parseCsvMap("mapDesert_Ground");
-const OBSTACLES: MapType = parseCsvMap("mapDesert_Objects");
-const TICK_RATE: number = 60;
-const BULLET_SPEED: number = 25;
-const TILE_SIZE: number = 32;
-
-const BUFFS: BuffType<BuffKey> = {
-    minorSwiftness: { movement: { speed: 0, multiplier: 0.2 }, duration: 30, name: "Minor Swiftness", tier: 1 },
-    mediumSwiftness: { movement: { speed: 0, multiplier: 0.4 }, duration: 22, name: "Medium Swiftness", tier: 2 },
-    majorSwiftness: { movement: { speed: 0, multiplier: 0.6 }, duration: 15, name: "Major Swiftness", tier: 3 },
-    rejuvenationMinor: { regen: { amount: 0, multiplier: 0.25 }, duration: 30, name: "Minor Rejuvenation", tier: 1 },
-    rejuvenationMedium: { regen: { amount: 0, multiplier: 0.50 }, duration: 22, name: "Medium Rejuvenation", tier: 2 },
-    rejuvenationMajor: { regen: { amount: 0, multiplier: 1 }, duration: 15, name: "Major Rejuvenation", tier: 3 },
-    minorFortification: { armor: { physical: 0, physicalMultiplier: 0.25 }, duration: 30, name: "Minor Fortification", tier: 1 },
-    mediumFortification: { armor: { physical: 0, physicalMultiplier: 0.50 }, duration: 22, name: "Medium Fortification", tier: 2 },
-    majorFortification: { armor: { physical: 0, physicalMultiplier: 1 }, duration: 15, name: "Major Fortification", tier: 3 },
-    minorWarding: { armor: { magic: 0, magicMultiplier: 0.25 }, duration: 30, name: "Minor Warding", tier: 1 },
-    mediumWarding: { armor: { magic: 0, magicMultiplier: 0.50 }, duration: 22, name: "Medium Warding", tier: 2 },
-    majorWarding: { armor: { magic: 0, magicMultiplier: 1 }, duration: 15, name: "Major Warding", tier: 3 },
-    minorStrengthening: { attack: { physical: 0, physicalMultiplier: 0.2 }, duration: 30, name: "Minor Strengthening", tier: 1 },
-    mediumStrengthening: { attack: { physical: 0, physicalMultiplier: 0.4 }, duration: 22, name: "Medium Strengthening", tier: 2 },
-    majorStrengthening: { attack: { physical: 0, physicalMultiplier: 0.6 }, duration: 15, name: "Major Strengthening", tier: 3 },
-    minorEnchantment: { attack: { magic: 0, magicMultiplier: 0.2 }, duration: 30, name: "Minor Enchantment", tier: 1 },
-    mediumEnchantment: { attack: { magic: 0, magicMultiplier: 0.4 }, duration: 22, name: "Medium Enchantment", tier: 2 },
-    majorEnchantment: { attack: { magic: 0, magicMultiplier: 0.6 }, duration: 15, name: "Major Enchantment", tier: 3 },
-    minorPrecision: { crit: { physicalRate: 0.05, magicRate: 0.05 }, duration: 30, name: "Minor Precision", tier: 1 },
-    mediumPrecision: { crit: { physicalRate: 0.1, magicRate: 0.1 }, duration: 22, name: "Medium Precision", tier: 2 },
-    majorPrecision: { crit: { physicalRate: 0.15, magicRate: 0.15 }, duration: 15, name: "Major Precision", tier: 3 },
-    minorDevastation: { crit: { physicalDamage: 0.1, magicDamage: 0.1 }, duration: 30, name: "Minor Devastation", tier: 1 },
-    mediumDevastation: { crit: { physicalDamage: 0.2, magicDamage: 0.2 }, duration: 22, name: "Medium Devastation", tier: 2 },
-    majorDevastation: { crit: { physicalDamage: 0.3, magicDamage: 0.3 }, duration: 15, name: "Major Devastation", tier: 3 },
-}
-
-type spawnRandomBuffType = () => { x: number, y: number, buff: BuffKey };
-const spawnRandomBuff: spawnRandomBuffType = () => {
-    const buffsList: BuffKey[] = Object.keys(BUFFS) as BuffKey[];
-    const randomBuff = buffsList[Math.floor(Math.random() * buffsList.length)];
-    const coords = generateRespawnCoords(OBSTACLES);
-
-    return {
-        x: coords.x,
-        y: coords.y,
-        buff: randomBuff,
-    };
-};
-
-class Player {
+class Player implements PlayerType {
     id: string;
     name: string;
     avatarIndex: number;
@@ -316,7 +271,7 @@ class Player {
     }
 }
 
-class Bullet {
+class Bullet implements BulletType {
     x: number; width: number;
     y: number; height: number;
     angle: number;
@@ -374,7 +329,7 @@ class Bullet {
     }
 }
 
-class Buff {
+class Buff implements BuffObjectType {
     x: number; y: number;
     type: BuffKey;
     duration: number;
@@ -398,9 +353,9 @@ class Buff {
             for (let innerProp in BUFFS[buff][outerProp]) {
                 if (!p.playerBuffStats[outerProp].hasOwnProperty(innerProp)) continue;
                 action === "apply" ?
-                //@ts-ignore
+                    //@ts-ignore
                     p.playerBuffStats[outerProp][innerProp] = parseFloat((p.playerBuffStats[outerProp][innerProp] + BUFFS[buff][outerProp][innerProp]).toFixed(2)) :
-                //@ts-ignore
+                    //@ts-ignore
                     p.playerBuffStats[outerProp][innerProp] = parseFloat((p.playerBuffStats[outerProp][innerProp] - BUFFS[buff][prop][innerProp]).toFixed(2));
             }
         }
